apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
